persistence fixes;

+++ common/src/main/java/com/simplegis/common/dto/OrganizationDto.java
@@ -5,7 +5,6 @@ import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -16,13 +15,13 @@ public class OrganizationDto implements Serializable {
 
     private static final long serialVersionUID = -880383329;
 
-    private BigInteger id;
+    private Long id;
     private String name;
     private Integer building;
     private Timestamp modified;
     private String www;
-    private BigInteger city;
-    private BigInteger street;
+    private Long city;
+    private Long street;
     private Integer scope;
     private List<PhoneDto> phones;
 
@@ -42,13 +41,13 @@ public class OrganizationDto implements Serializable {
     }
 
     public OrganizationDto(
-            BigInteger id,
+            Long id,
             String name,
             Integer building,
             Timestamp modified,
             String www,
-            BigInteger city,
-            BigInteger street,
+            Long city,
+            Long street,
             Integer scope,
             List<PhoneDto> phones
     ) {
@@ -67,11 +66,11 @@ public class OrganizationDto implements Serializable {
         return serialVersionUID;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -107,19 +106,19 @@ public class OrganizationDto implements Serializable {
         this.www = www;
     }
 
-    public BigInteger getCity() {
+    public Long getCity() {
         return city;
     }
 
-    public void setCity(BigInteger city) {
+    public void setCity(Long city) {
         this.city = city;
     }
 
-    public BigInteger getStreet() {
+    public Long getStreet() {
         return street;
     }
 
-    public void setStreet(BigInteger street) {
+    public void setStreet(Long street) {
         this.street = street;
     }
